% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cologne.R
\name{cologne}
\alias{cologne}
\title{Cologne Phonetic Name Coding}
\usage{
cologne(word, maxCodeLen = NULL, clean = TRUE)
}
\arguments{
\item{word}{string or vector of strings to encode}

\item{maxCodeLen}{maximum length of the resulting encodings, in characters}

\item{clean}{if \code{TRUE}, return \code{NA} for unknown alphabetical characters}
}
\value{
the Cologne encoded character vector
}
\description{
The Cologne phonetic name coding procedure.
}
\details{
\code{maxCodeLen = NULL} preserves the algorithm's historical unbounded
output. An explicit nonnegative whole-number value truncates each code to
at most that many characters.

The \code{cologne} algorithm is only defined for inputs over the
standard English alphabet, \emph{i.e.}, "A-Z," "Ä," "Ö," "Ü," and
"ß." Non-alphabetical characters are removed from the string in a
locale-dependent fashion.  This strips spaces, hyphens, and numbers.
Other letters, such as "ç," may be permissible in the current locale
but are unknown to \code{cologne}.  For inputs outside of its known
range, the output is undefined and \code{NA} is returned and a
\code{warning} is issued.  If \code{clean} is \code{FALSE},
\code{cologne} attempts to process the strings.  The default is
\code{TRUE}.
}
\examples{
cologne("William")
cologne(c("Peter", "Peady"))
cologne("Stevenson", maxCodeLen = 8)

}
\references{
James P. Howard, II, "Phonetic Spelling Algorithm Implementations
for R," \emph{Journal of Statistical Software}, vol. 95, no. 8,
(2020), p. 1--21, <10.18637/jss.v095.i08>.

Hans Joachim Postel. "Die Koelner Phonetik. Ein Verfahren zur
Identifizierung von Personennamen auf der Grundlage der
Gestaltanalyse."  \emph{IBM-Nachrichten} 19. Jahrgang, 1969,
p. 925-931.
}
\seealso{
Other phonics:
\code{\link[=caverphone]{caverphone()}},
\code{\link[=lein]{lein()}},
\code{\link[=metaphone]{metaphone()}},
\code{\link[=mra_encode]{mra_encode()}},
\code{\link[=nysiis]{nysiis()}},
\code{\link[=onca]{onca()}},
\code{\link[=phonex]{phonex()}},
\code{\link[=phonics]{phonics()}},
\code{\link[=rogerroot]{rogerroot()}},
\code{\link[=soundex]{soundex()}},
\code{\link[=statcan]{statcan()}}
}
\concept{phonics}
