% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphone.R
\name{metaphone}
\alias{metaphone}
\title{Generate phonetic versions of strings with Metaphone}
\usage{
metaphone(word, maxCodeLen = 10L, clean = TRUE)
}
\arguments{
\item{word}{string or vector of strings to encode}

\item{maxCodeLen}{maximum length of the resulting encodings, in characters}

\item{clean}{if \code{TRUE}, return \code{NA} for unknown alphabetical characters}
}
\value{
a character vector containing the metaphones of \code{word},
or an NA if the \code{word} value is NA
}
\description{
The function \code{metaphone} phonetically encodes the
given string using the metaphone algorithm.
}
\details{
Metaphone has accumulated incompatible implementation variants. This
function follows the traditional rules from Philips' 1990 algorithm; it is
therefore not expected to reproduce PHP or Apache Commons Codec for every
input.

The variable \code{maxCodeLen} is the limit on how long the returned
metaphone should be.

The \code{metaphone} algorithm is only defined for inputs over the
standard English alphabet, \emph{i.e.}, "A-Z.". Non-alphabetical
characters are removed from the string in a locale-dependent fashion.
This strips spaces, hyphens, and numbers.  Other letters, such as
"Ü," may be permissible in the current locale but are unknown to
\code{metaphone}.  For inputs outside of its known range, the output
is undefined and \code{NA} is returned and a \code{warning} this
thrown.  If \code{clean} is \code{FALSE}, \code{metaphone} attempts
to process the strings.  The default is \code{TRUE}.
}
\examples{
metaphone("wheel")
metaphone(c("school", "benji"))

}
\references{
James P. Howard, II, "Phonetic Spelling Algorithm Implementations
for R," \emph{Journal of Statistical Software}, vol. 95, no. 8,
(2020), p. 1--21, <10.18637/jss.v095.i08>.

Lawrence Philips, "Hanging on the Metaphone," \emph{Computer Language},
vol. 7, no. 12 (1990), p. 38--44.
}
\seealso{
Other phonics:
\code{\link[=caverphone]{caverphone()}},
\code{\link[=cologne]{cologne()}},
\code{\link[=lein]{lein()}},
\code{\link[=mra_encode]{mra_encode()}},
\code{\link[=nysiis]{nysiis()}},
\code{\link[=onca]{onca()}},
\code{\link[=phonex]{phonex()}},
\code{\link[=phonics]{phonics()}},
\code{\link[=rogerroot]{rogerroot()}},
\code{\link[=soundex]{soundex()}},
\code{\link[=statcan]{statcan()}}
}
\concept{phonics}
