% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonics-package.R
\docType{package}
\name{phonics-package}
\alias{phonics-package}
\title{Phonetic spelling algorithms}
\description{
The pkg{phonics} package provides phonetic encoders for English, German,
and French names, together with Match Rating Approach comparison and a
convenience dispatcher. See code{link{phonics}} for the complete method
list and code{vignette("phonics")} for an overview.
}
\references{
Howard, J. P., II (2020). "Phonetic Spelling Algorithm Implementations
for R." emph{Journal of Statistical Software}, 95(8), 1--21.
doi{10.18637/jss.v095.i08}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://phonics.jameshoward.us}
  \item \url{https://github.com/k3jph/phonics-in-r}
  \item Report bugs at \url{https://github.com/k3jph/phonics-in-r/issues}
}

}
\author{
\strong{Maintainer}: James Howard \email{jh@jameshoward.us} (\href{https://orcid.org/0000-0003-4530-1547}{ORCID})

Authors:
\itemize{
  \item James Howard \email{jh@jameshoward.us} (\href{https://orcid.org/0000-0003-4530-1547}{ORCID})
}

Other contributors:
\itemize{
  \item Kyle Haynes [contributor]
  \item Amanda Hood [contributor]
  \item Os Keyes [contributor]
}

}
\keyword{internal}
