% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonics.R
\name{phonics}
\alias{phonics}
\title{Phonetic Spelling Algorithms}
\usage{
phonics(word, method, clean = TRUE)
}
\arguments{
\item{word}{A character vector to encode.}

\item{method}{A character vector naming the methods to apply.}

\item{clean}{If \code{TRUE}, warn and return \code{NA} for inputs containing
characters outside an encoder's supported alphabet. If \code{FALSE},
unsupported characters are discarded before encoding.}
}
\value{
A data frame whose first column is \code{word}, followed by one character
column for each recognized method.
}
\description{
Apply one or more phonetic encoders to a character vector and return their
codes together in a data frame.
}
\details{
Available methods are \code{"caverphone"},
\code{"caverphone.modified"}, \code{"cologne"}, \code{"lein"},
\code{"metaphone"}, \code{"nysiis"}, \code{"nysiis.modified"},
\code{"onca"}, \code{"onca.modified"}, \code{"onca.refined"},
\code{"onca.modified.refined"}, \code{"phonex"},
\code{"rogerroot"}, \code{"soundex"}, \code{"soundex.refined"}, and
\code{"statcan"}. Unknown method names produce a warning and no result
column.
}
\examples{
phonics(c("Peter", "Peady"), c("soundex", "soundex.refined"))

}
\references{
Howard, J. P., II (2020). "Phonetic Spelling Algorithm Implementations
for R." \emph{Journal of Statistical Software}, 95(8), 1--21.
\doi{10.18637/jss.v095.i08}.
}
\seealso{
Other phonics:
\code{\link[=caverphone]{caverphone()}},
\code{\link[=cologne]{cologne()}},
\code{\link[=lein]{lein()}},
\code{\link[=metaphone]{metaphone()}},
\code{\link[=mra_encode]{mra_encode()}},
\code{\link[=nysiis]{nysiis()}},
\code{\link[=onca]{onca()}},
\code{\link[=phonex]{phonex()}},
\code{\link[=rogerroot]{rogerroot()}},
\code{\link[=soundex]{soundex()}},
\code{\link[=statcan]{statcan()}}
}
\concept{phonics}
