% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundex.R
\name{soundex}
\alias{soundex}
\alias{refinedSoundex}
\title{Soundex}
\usage{
soundex(word, maxCodeLen = 4L, clean = TRUE)

refinedSoundex(word, maxCodeLen = 10L, clean = TRUE)
}
\arguments{
\item{word}{string or vector of strings to encode}

\item{maxCodeLen}{maximum length of the resulting encodings, in characters}

\item{clean}{if \code{TRUE}, return \code{NA} for unknown alphabetical characters}
}
\value{
soundex encoded character vector
}
\description{
The Soundex phonetic algorithms
}
\details{
The function \code{soundex} phonetically encodes the given
string using the U.S. census Soundex rules. The function
\code{refinedSoundex} implements the U.S.-English mapping and duplicate
handling from Apache Commons Codec's Refined Soundex.

The variable \code{maxCodeLen} is the limit on how long the returned
soundex should be.

The \code{soundex} and \code{refinedSoundex} algorithms are only
defined for inputs over the standard English alphabet, \emph{i.e.},
"A-Z." Non-alphabetical characters are removed from the string in a
locale-dependent fashion.  This strips spaces, hyphens, and numbers.
Other letters, such as "Ü," may be permissible in the current locale
but are unknown to \code{soundex} and \code{refinedSoundex}.  For
inputs outside of its known range, the output is undefined and
\code{NA} is returned and a \code{warning} is issued.  If
\code{clean} is \code{FALSE}, \code{soundex} and
\code{refinedSoundex} attempts to process the strings.  The default
is \code{TRUE}.
}
\section{Caveats}{

The \code{soundex} and \code{refinedSoundex} algorithms are only
defined for inputs over the standard English alphabet, \emph{i.e.},
"A-Z." For inputs outside this range, the output is undefined.
}

\examples{
soundex("wheel")
soundex(c("school", "benji"))

}
\references{
Charles P. Bourne and Donald F. Ford, "A study of methods for
systematically abbreviating English words and names," \emph{Journal
of the ACM}, vol. 8, no. 4 (1961), p. 538-552.

James P. Howard, II, "Phonetic Spelling Algorithm Implementations
for R," \emph{Journal of Statistical Software}, vol. 95, no. 8,
(2020), p. 1--21, <10.18637/jss.v095.i08>.

Howard B. Newcombe, James M. Kennedy, "Record linkage: making
maximum use of the discriminating power of identifying information,"
\emph{Communications of the ACM}, vol. 5, no. 11 (1962), p. 563-566.

United States National Archives and Records Administration,
"The Soundex Indexing System."

Apache Software Foundation, \emph{Apache Commons Codec:
RefinedSoundex}.
}
\seealso{
Other phonics:
\code{\link[=caverphone]{caverphone()}},
\code{\link[=cologne]{cologne()}},
\code{\link[=lein]{lein()}},
\code{\link[=metaphone]{metaphone()}},
\code{\link[=mra_encode]{mra_encode()}},
\code{\link[=nysiis]{nysiis()}},
\code{\link[=onca]{onca()}},
\code{\link[=phonex]{phonex()}},
\code{\link[=phonics]{phonics()}},
\code{\link[=rogerroot]{rogerroot()}},
\code{\link[=statcan]{statcan()}}
}
\concept{phonics}
