% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metaphone}
\alias{metaphone}
\title{metaphone}
\usage{
metaphone(word, maxCodeLen = 10L, traditional = TRUE)
}
\arguments{
\item{word}{character vector of strings to encode}

\item{maxCodeLen}{maximum length of the resulting encodings, in characters}

\item{traditional}{use traditional Metaphone or not}
}
\value{
metaphone encoded character vector
}
\description{
Encode a string with the Metaphone phonentic algorithm
}
\details{
The function \code{metaphone} phonentically encodes the '
given string using the metaphone algorithm.  There is some discrepency
with respect to how the metaphone algorithm actually works.  For '
instance, there is a version in the Java Apache Commons library.  '
There is a version provided within PHP.  These do not provide the same
results.  On the questionable theory that the implementation in PHP
is probably more well known, this code should match it in output.

This implementation is based on a Javascript implementation which is
itself based on the PHP internal implementation.
}
\examples{
metaphone("wheel")
metaphone(c("school", "benji"))
}

