% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonics.R
\docType{package}
\name{phonics-package}
\alias{phonics}
\alias{phonics-package}
\title{Phonetic Spelling Algorithms}
\description{
Encode words phonetic spelling algorithms.
}
\details{
The \code{phonics} package for \code{R} is designed to provide a
variety of phonetic indexing algorithms in common and not-so-common
use today.  The algorithms generally reduce a string to a symbolic
representation approximating the sound made by pronouncing the
string.  They can be used to match names, words, and as a proxy for
assorted string distance algorithms.  The algorithm reduces a string
to a symbolic representation approximating the sound.  It can be used
to match names, words, and as a proxy for assorted string distance
algorithms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jameshoward.us/software/phonics/}
  \item \url{https://github.com/howardjp/phonics}
  \item Report bugs at \url{https://github.com/howardjp/phonics/issues}
}

}
\author{
James P. Howard, II <jh@jameshoward.us>
}
