% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonics.R
\docType{package}
\name{phonics-package}
\alias{phonics}
\alias{phonics-package}
\title{Phonetic Spelling Algorithms}
\description{
Encode words with English-language Metaphone
}
\details{
The phonics package provides an implementation of the Metaphone
phonetic algorithm in R.  The algorithm reduces a string to a
symbolic representation approximating the sound.  It can be used to
match names, words, and as a proxy for assorted string distance
algorithms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jameshoward.us/software/phonics/}
  \item \url{https://github.com/howardjp/phonics}
  \item Report bugs at \url{https://github.com/howardjp/phonics/issues}
}

}
\author{
James P. Howard, II <jh@jameshoward.us>
}
