% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonics.R
\name{phonics}
\alias{phonics}
\title{Phonetic Spelling Algorithms}
\usage{
phonics(word, method, clean = TRUE)
}
\arguments{
\item{word}{string or vector of strings to encode}

\item{method}{vector of method names to use}

\item{clean}{if \code{TRUE}, return \code{NA} for unknown alphabetical characters}
}
\value{
Returns a data frame containing the phonetic spellings of the input
for each method applied.
}
\description{
The \code{phonics} package for \code{R} is designed to provide a
variety of phonetic indexing algorithms in common and not-so-common
use today.  The algorithms generally reduce a string to a symbolic
representation approximating the sound made by pronouncing the
string.  They can be used to match names, strings, and as a proxy for
assorted string distance algorithms.  The algorithm reduces a string
to a symbolic representation approximating the sound.  It can be used
to match names, strings, and as a proxy for assorted string distance
algorithms.
}
\details{
The \code{phonics} package for \code{R} is designed to provide a
variety of phonetic indexing algorithms in common and not-so-common
use today.  The algorithms generally reduce a string to a symbolic
representation approximating the sound made by pronouncing the
string.  They can be used to match names, strings, and as a proxy for
assorted string distance algorithms.  The algorithm reduces a string
to a symbolic representation approximating the sound.  It can be used
to match names, strings, and as a proxy for assorted string distance
algorithms.

The variable \code{word} is the name to be encoded.  The variable
\code{maxCodeLen} is \emph{not} supported in this algorithm encoder
because the algorithm itself is dependent upon its six-character
length.  The variables \code{x} and \code{y} are MRA-encoded and are
compared to each other using the MRA comparison specification.

Different phonetic algorithm are only defined for inputs over the
limited alphabets, Non-alphabetical characters are removed from the
string in a locale-dependent fashion.  This strips spaces, hyphens,
and numbers.  For inputs outside of its known range, the output is
undefined and \code{NA} is returned and a \code{warning} this thrown.
If \code{clean} is \code{FALSE}, \code{phonics} attempts to process
the strings.  The default is \code{TRUE}.
}
\examples{
phonics(c("Peter", "Peady"), c("soundex", "soundex.refined"))

}
\seealso{
Other phonics: \code{\link{caverphone}},
  \code{\link{cologne}}, \code{\link{lein}},
  \code{\link{metaphone}}, \code{\link{mra_encode}},
  \code{\link{nysiis}}, \code{\link{onca}},
  \code{\link{phonex}}, \code{\link{rogerroot}},
  \code{\link{soundex}}, \code{\link{statcan}}
}
\concept{phonics}
