% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogerroot.R
\name{rogerroot}
\alias{rogerroot}
\title{Roger Root Name Coding Procedure}
\usage{
rogerroot(word, maxCodeLen = 5, ignoreNonAlpha = FALSE)
}
\arguments{
\item{word}{string or vector of strings to encode}

\item{maxCodeLen}{maximum length of the resulting encodings, in characters}

\item{ignoreNonAlpha}{if \code{TRUE}, ignore non-alphabetic chracters}
}
\value{
the Roger Root encoded character vector
}
\description{
Provides the Roger Root name coding system
}
\details{
The \code{rogerroot} function phentically encodes the given string
using the Roger Root algorithm.  The variable \code{word} is a string
or vector of strings to encode.

The variable \code{maxCodeLen} is the limit on how long the returned
code should be.  The default is 5.

The \code{rogerroot} algorithm is only defined for inputs over the
standard English alphabet, \emph{i.e.}, "A-Z." For inputs outside
this range, the output is undefined and \code{NA} is returned.  If
\code{ignoreNonAlpha} is \code{TRUE}, \code{rogerroot} attempts to
process the strings.
}
\examples{
rogerroot("William")
rogerroot(c("Peter", "Peady"))
rogerroot("Stevenson")

}
\references{
Robert L. Taft, \emph{Name search techniques}, Bureau of Systems
Development, Albany, New York, 1970.
}
\seealso{
Other phonics: \code{\link{caverphone}},
  \code{\link{cologne}}, \code{\link{lein}},
  \code{\link{metaphone}}, \code{\link{mra_encode}},
  \code{\link{nysiis}}, \code{\link{onca}},
  \code{\link{phonex}}, \code{\link{soundex}},
  \code{\link{statcan}}
}
\concept{phonics}
