% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rogerroot.R
\name{rogerroot}
\alias{rogerroot}
\title{Roger Root Name Coding Procedure}
\usage{
rogerroot(word, maxCodeLen = 5)
}
\arguments{
\item{word}{character or character vector or wordings to encode}

\item{maxCodeLen}{maximum length of the resulting encodings, in characters}
}
\value{
The \code{mra_encode} function returns match rating approach
encoded character vector.  The \code{mra_compare} returns a boolean
vector which is \code{TRUE} if \code{x} and \code{y} pass the MRA
comparison test.
}
\description{
Provides the Roger Root name coding system
}
\details{
The variable \code{word} is the name to be encoded.  The variable
\code{maxCodeLen} is \emph{not} supported in this algorithm encoder
because the algorithm itself is dependent upon its six-character
length.  The variables \code{x} and \code{y} are MRA-encoded and are
compared to each other using the MRA comparison specification.
}
\examples{
rogerroot("William")
rogerroot(c("Peter", "Peady"))
rogerroot("Stevenson")

}
\references{
Robert L. Taft, \emph{Name search techniques}, Bureau of Systems
Development, Albany, New York, 1970.
}
\seealso{
Other phonics: \code{\link{caverphone}},
  \code{\link{lein}}, \code{\link{metaphone}},
  \code{\link{mra_encode}}, \code{\link{nysiis}},
  \code{\link{soundex}}, \code{\link{statcan}}
}

